requirements.txt
fastapi
uvicorn
google-genai
pinecone-client
tenacity
python-dotenv
